+++ docs/changelog.md
@@ -2,6 +2,14 @@
 
 ## v2
 
+### 2.3.x
+
+- Added support for the light/colour sensor in the v2 Sense HAT
+
+### 2.2.0
+
+- Added new stick interface for the joystick
+
 ### 2.1.0
 
 - Added gamma, low light and other properties